backend/requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.7
fastapi==0.115.2
firebase_admin==6.5.0
fitz==0.0.1.dev2
google_api_python_client==2.145.0
google_auth_oauthlib==1.2.1
google_cloud_aiplatform==1.67.1
groq==0.11.0
ipython==8.12.3
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.10
langchain_google_genai==2.0.1
langchain_google_vertexai==2.0.5
langchain_groq==0.2.0
langgraph==0.2.35
matplotlib==3.9.2
networkx==3.3
Pillow==10.4.0
pinecone==5.3.1
protobuf==5.28.2
pydantic==2.9.2
PyPDF2==3.0.1
pysbd==0.3.4
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
pytz==2024.1
pywhatkit==5.4
Requests==2.32.3
semantic_router==0.0.72
spacy==3.7.6
tqdm==4.66.5
trafilatura==1.12.2
youtube_transcript_api==0.6.2
